KAFKA-18263; Group lock must be acquired when reverting static membership rejoin (#18207)
When a static member rejoins the group, the group state is rewritten to the partition in order to persist the change. If the write fails, the change is reverted. However, this is done without acquiring the group lock.

This is only try in the old group coordinator. The new one does not suffer from this issue.

Reviewers: Jeff Kim <jeff.kim@confluent.io>